server/requirement.txt
fastapi 
uvicorn 
python-dotenv 
google-generativeai